requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
cython==3.1.2
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
formulaic==1.2.0
idna==3.10
interface-meta==1.3.0
ipykernel==6.30.1
ipython==9.4.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-pygments==0.3.0
linearmodels==6.1
markupsafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
mypy-extensions==1.1.0
narwhals==2.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pygments==2.19.2
pyhdfe==0.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.27.0
scipy==1.16.1
setuptools==80.9.0
setuptools-scm==8.3.1
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.5
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.17.2
yarg==0.1.9
